:recycle: Refactor cart endpoint
Don't put data in enveloppe
