+++ src/brouwers/shop/api/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
 
-from .views import IdealBanksView
-from .viewsets import CartViewSet
+from .views import CartView, IdealBanksView
 
 urlpatterns = [
-    path("cart/", CartViewSet.as_view(), name="cart-detail"),
+    path("cart/", CartView.as_view(), name="cart-detail"),
     path("ideal_banks/", IdealBanksView.as_view(), name="ideal-banks"),
 ]
