+++ src/brouwers/shop/tests/test_api.py
@@ -4,7 +4,7 @@ from django.core.cache import cache
 from django.urls import reverse
 
 from rest_framework import status
-from rest_framework.test import APITransactionTestCase
+from rest_framework.test import APITestCase
 
 from brouwers.users.tests.factories import UserFactory
 
@@ -13,12 +13,14 @@ from ..models import Cart
 from .factories import CartFactory, CartProductFactory, ProductFactory
 
 
-class ProductApiTest(APITransactionTestCase):
+class ProductApiTest(APITestCase):
     """
     Test that CRUD operations for products work correctly
     """
 
     def setUp(self):
+        super().setUp()
+
         cache.clear()
         self.user = UserFactory.create()
         self.client.force_authenticate(user=self.user)
@@ -26,19 +28,23 @@ class ProductApiTest(APITransactionTestCase):
 
     def test_get_product(self):
         product = ProductFactory.create()
+
         response = self.client.get(reverse("api:product-detail", args=[product.pk]))
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data["id"], product.pk)
         self.assertEqual(response.data["name"], product.name)
         self.assertEqual(response.data["brand"], product.brand.id)
 
 
-class CartApiTest(APITransactionTestCase):
+class CartApiTest(APITestCase):
     """
     Test that CRUD operations for cart work correctly
     """
 
     def setUp(self):
+        super().setUp()
+
         cache.clear()
         self.user = UserFactory.create()
         self.client.force_authenticate(user=self.user)
@@ -46,44 +52,54 @@ class CartApiTest(APITransactionTestCase):
 
     def test_get_cart(self):
         cart = CartFactory.create(user=self.user)
+
         response = self.client.get(reverse("api:cart-detail"))
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["cart"]["id"], cart.id)
-        self.assertEqual(response.data["cart"]["status"], cart.status)
+        self.assertEqual(response.data["id"], cart.id)
+        self.assertEqual(response.data["status"], cart.status)
 
         # Should return the last open cart
         open_cart = CartFactory.create(user=self.user, status=CartStatuses.open)
         closed_cart = CartFactory.create(user=self.user, status=CartStatuses.paid)
+
         response = self.client.get(reverse("api:cart-detail"))
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["cart"]["id"], open_cart.id)
-        self.assertNotEqual(response.data["cart"]["id"], closed_cart.id)
+        self.assertEqual(response.data["id"], open_cart.id)
+        self.assertNotEqual(response.data["id"], closed_cart.id)
 
         # User shouldn't be able see other users' carts
         user2 = UserFactory.create()
         self.client.force_authenticate(user=user2)
         cart2 = CartFactory.create(user=user2)
+
         response = self.client.get(reverse("api:cart-detail"))
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(response.data["cart"]["id"], cart.id)
+        self.assertNotEqual(response.data["id"], cart.id)
 
         # Anon user should get a new cart
         self.client.logout()
+
         response = self.client.get(reverse("api:cart-detail"))
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        cart = response.data["cart"]
+        cart = response.data
 
         # Should get the same cart after the second request
         response2 = self.client.get(reverse("api:cart-detail"))
         self.assertEqual(response2.status_code, status.HTTP_200_OK)
-        self.assertEqual(response2.data["cart"]["id"], cart["id"])
+        self.assertEqual(response2.data["id"], cart["id"])
 
     def test_get_cart_product(self):
         cart = CartFactory.create(user=self.user)
         cart_product = CartProductFactory.create(cart=cart)
+
         response = self.client.get(
             reverse("api:cartproduct-detail", args=[cart_product.pk])
         )
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data["id"], cart_product.pk)
         self.assertEqual(response.data["amount"], cart_product.amount)
@@ -93,16 +109,20 @@ class CartApiTest(APITransactionTestCase):
         user2 = UserFactory.create()
         cart2 = CartFactory.create(user=user2)
         cart_product2 = CartProductFactory.create(cart=cart2)
+
         response = self.client.get(
             reverse("api:cartproduct-detail", args=[cart_product2.pk])
         )
+
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_get_cart_products_by_cart(self):
         cart = CartFactory.create(user=self.user)
         CartProductFactory.create(cart=cart, amount=2)
         CartProductFactory.create(cart=cart, amount=20)
+
         response = self.client.get(reverse("api:cartproduct-list"), {"cart": cart.pk})
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(len(response.data), 2)
 
@@ -110,7 +130,9 @@ class CartApiTest(APITransactionTestCase):
         cart = CartFactory.create(user=self.user)
         product = ProductFactory.create()
         data = {"product": product.id, "cart": cart.id, "amount": 13}
+
         response = self.client.post(reverse("api:cartproduct-list"), data)
+
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         cart = Cart.objects.get(id=cart.id)
         self.assertEqual(cart.products.count(), 1)
@@ -130,8 +152,10 @@ class CartApiTest(APITransactionTestCase):
         cart = CartFactory.create(user=self.user)
         product = ProductFactory.create()
         data = {"product": product.id, "cart": cart.id, "amount": 1}
+
         self.client.post(reverse("api:cartproduct-list"), data)
         self.client.post(reverse("api:cartproduct-list"), data)
+
         cart = Cart.objects.get(id=cart.id)
         self.assertEqual(cart.products.count(), 1)
 
@@ -140,9 +164,11 @@ class CartApiTest(APITransactionTestCase):
         cart_product = CartProductFactory.create(
             product=product, amount=4, cart__user=self.user
         )
+
         response = self.client.get(
             reverse("api:cartproduct-detail", args=[cart_product.pk])
         )
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data["total"], Decimal("4.60"))
 
@@ -154,6 +180,8 @@ class CartApiTest(APITransactionTestCase):
         cart_product2 = CartProductFactory.create(
             product=product2, amount=40, cart=cart
         )
+
         response = self.client.get(reverse("api:cart-detail"))
+
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["cart"]["total"], Decimal("10.60"))
+        self.assertEqual(response.data["total"], Decimal("10.60"))
